requirements.txt
asgiref==3.2.7
attrs==19.3.0
autoflake==1.4
autopep8==1.5.5
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
docformatter==1.4
idna==2.9
importlib-metadata==1.6.0
isort==5.7.0
more-itertools==8.2.0
packaging==20.3
Pillow==8.1.1
pluggy==0.13.1
py==1.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
pyformat==0.7
PyJWT==2.0.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
python-dotenv==0.15.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.2
unify==0.5
untokenize==0.1.1
urllib3==1.25.9
wcwidth==0.1.9
zipp==3.1.0
